Avoid creating garbage by using the same instances of sentinel extensions and frame consumers for incoming and outgoig messages. Created OutgoingSentinelExtension and IncomingSentinelExtension.
